Added a default message for the invalid commands (#55)
* Update bot.js

Added a default message for the commands which aren't valid

* Added a default message for the invalid commands #53

Issue #53 ,Bot was not showing any response for unknown triggers. Fixed it by adding a default message whenever a invalid command is triggered.

* Bot not responding to unknown Commands #53

Issue #53, Bot was not showing any response for unknown triggers. Fixed it by adding a default message whenever a invalid command is triggered.

* Update bot.js

* Removed unwanted return statement